requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.71.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
loguru==0.5.3
loguru-logging-intercept==0.1.1
Mako==1.1.6
MarkupSafe==2.0.1
orjson==3.6.5
packaging==21.3
passlib==1.7.4
pendulum==2.1.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytzdata==2020.1
PyYAML==5.4.1
ratelimit==2.2.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvicorn-loguru-integration==0.3.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
